Rework loot table view
The current view only displays one item per line, which becomes a problem when loot tables are large. As loot tables are mainly about what items are in them, this PR simplifies the view to just show an inline list of slots, just like LU's inventory does. As for the other columns, `id` never seems to be used in the game and is probably an auto-added primary key. `missionDrop` is indicated by highlighting the slot. `sortPriority` is not displayed directly, but is used to order the items. Sorting is also moved from JS to CSS, as it is only used in presentation. As `sortPriority` is also only used by LU to order items, without the precise value being of much importance, not displaying it directly seems like an acceptable tradeoff for a much more compact representation of loot tables.